convert 1e6*hertz to long
git-svn-id: https://svn.atnf.csiro.au/difx@1773 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/branches/pcal/src/configuration.cpp
@@ -1106,7 +1106,7 @@ bool Configuration::processDatastreamTable(ifstream * input)
           tonefreq += dsdata->phasecalintervalmhz;
         for(int k=0;k<datastreamtable[i].numrecordedfreqpcaltones[j];k++)
           datastreamtable[i].recordedfreqpcaltonefreqs[j][k] = tonefreq + k*dsdata->phasecalintervalmhz;
-        dsdata->recordedfreqpcaloffsetshz[j] = 1e6*datastreamtable[i].recordedfreqpcaltonefreqs[j][0] - 1e6*lofreq;
+        dsdata->recordedfreqpcaloffsetshz[j] = long(1e6*datastreamtable[i].recordedfreqpcaltonefreqs[j][0]) - long(1e6*lofreq);
       }
     }
     datastreamtable[i].tcpwindowsizekb = 0;
